+++ R/get_extract_vd_chunked.R
@@ -36,6 +36,29 @@ get_extract_vd_chunked <- function(ids, chunksize = 20, cols = NA, returnunique 
     basereq <- vb_basereq()
   }
 
+  if (length(ids) > 10) {
+    # Preflight ssl check
+    status <- tryCatch({
+      preflight_test <- basereq %>% req_perform()  # nolint: object_usage_linter
+      list("err_code" = 0, "err_obj" = NULL)
+    }, error = function(e) {
+      list("err_code" = 1, "err_obj" = e)
+    })
+
+    if (status$err_code == 1) {
+      curl_err <- get_curl_err(status$err_obj)
+      if (grepl("SSL certificate problem: unable to get local issuer certificate", curl_err)) {
+        cat("\n")
+        cli_alert_danger("Could not verify SSL certificate.")
+        cli::cli_text("You may have success running {.fn set_ohvbd_compat} and then trying again.")
+        cat("\n")
+        cli_abort("SSL certificate problem: unable to get local issuer certificate")
+      } else {
+        cli_abort("Preflight found unknown error: {.val {curl_err}}")
+      }
+    }
+  }
+
   # Get and extract vt data by ID in chunks (to save memory)
 
   # Split into chunks
@@ -49,7 +72,7 @@ get_extract_vd_chunked <- function(ids, chunksize = 20, cols = NA, returnunique 
       extract_vd(cols = cols, returnunique = returnunique)
   })
 
-  out_df <- rbindlist(out_list, fill = TRUE)
+  out_df <- suppressWarnings(rbindlist(out_list, fill = TRUE))
 
   out_final <- as.data.frame(out_df)
   attr(out_final, "db") <- "vd"