+++ R/search_vd.R
@@ -33,16 +33,22 @@ search_vd <- function(keywords, basereq = NA) {
       req_url_query("format" = "json") %>%
       req_url_query("keywords" = keywords, .multi = space_collapse) %>%
       req_perform()
-    list("resp" = resp, "err_code" = 0)
+    list("resp" = resp, "err_code" = 0, "err_obj" = NULL)
   }, error = function(e) {
     # Get the last response instead
-    list("resp" = last_response(), "err_code" = 1)
+    list("resp" = last_response(), "err_code" = 1, "err_obj" = e)
   })
 
   if (resplist$err_code == 1) {
+    if (grepl("SSL certificate problem: unable to get local issuer certificate", resplist$err_obj$parent$message)) {
+      cat("\n")
+      cli_alert_danger("Could not verify SSL certificate.")
+      cli::cli_text("You may have success running {.fn set_ohvbd_compat} and then trying again.")
+      cat("\n")
+      cli_abort("SSL certificate problem: unable to get local issuer certificate")
+    }
     cli_abort("No records found for {.val {keywords}}")
   }
-
   body <- resplist$resp %>% resp_body_json()
 
   if (length(body) > 2) {